Split courses into general and admin
